# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: 'SAKA Marketplace API'
  version: 1.0.0
  description: |-
    Multi-vertical marketplace API for Tanzania.

    **Errors.** Every failure returns the same envelope:

        { "error": { "code": "...", "message": "...", "details": {...}, "request_id": "..." } }

    Clients should switch on `error.code`, never on `message`. Validation failures (422) additionally carry Laravel's `errors` map.

    **Correlation.** Every response carries `X-Request-Id`; an inbound value is honoured when it matches `^[A-Za-z0-9_-]{8,64}$`.

    **Authentication.** Bearer tokens issued by `/auth/login`. Tokens are unscoped; authorization is enforced by server-side policies.

    **Not found vs forbidden.** Resources the caller may not see return 404, never 403, so existence is not disclosed.
  contact:
    name: 'SAKA Engineering'
  license:
    name: Proprietary
servers:
  -
    url: 'https://api.saka.co.tz/api/v1'
    description: Production
  -
    url: 'http://localhost:8000/api/v1'
    description: Local
...
paths:
...
  /account/reviews:
    get:
      operationId: api_v1_account_reviews_mine
      summary: 'Reviews written by the current user'
      tags:
        - Reviews
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  -
                    $ref: '#/components/schemas/PaginatedEnvelope'
                  -
                    type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Review' } }
        401:
          $ref: '#/components/responses/Unauthenticated'
        403:
          $ref: '#/components/responses/Forbidden'
        429:
          $ref: '#/components/responses/RateLimited'
        500:
          $ref: '#/components/responses/ServerError'
      description: |-


        Rate limit: `api-authenticated`.
      security:
        -
          bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'Token from `/auth/login`, `/auth/register` or `/auth/oauth/google`.'
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            request_id:
              type: string
          required:
            - code
            - message
      required:
        - error
...
    PaginatedEnvelope:
      type: object
      properties:
        data:
          type: array
          items: []
        links:
          type: object
          properties:
            first:
              type:
                - string
                - 'null'
            last:
              type:
                - string
                - 'null'
            prev:
              type:
                - string
                - 'null'
            next:
              type:
                - string
                - 'null'
        meta:
          type: object
          properties:
            current_page:
              type: integer
            per_page:
              type: integer
            total:
              type: integer
            last_page:
              type: integer
...
    Review:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type:
            - string
            - 'null'
        body:
          type:
            - string
            - 'null'
        status:
          type: string
          enum:
            - pending
            - approved
            - rejected
        helpful_count:
          type: integer
        reply:
          oneOf:
            -
              type: object
              properties:
                body:
                  type: string
                replied_at:
                  type: string
            -
              type: 'null'
        reviewer:
          type: object
          properties:
            uuid:
              type: string
            name:
              type: string
        created_at:
          type: string
          format: date-time
...
  responses:
    Unauthenticated:
      description: 'Missing or invalid token (`UNAUTHENTICATED`).'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 'Authenticated but not permitted (`FORBIDDEN`).'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: 'Rate limit exceeded (`RATE_LIMITED`). See `Retry-After`.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: 'Unexpected error (`SERVER_ERROR`). No internal detail is disclosed.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'